requirements.txt
flask>=3.0.2
flask-cors>=4.0.0
flask-jwt-extended>=4.6.0
flask-limiter>=3.6.0
flask-sqlalchemy>=3.1.1
sqlalchemy>=2.0.25
langchain>=0.2.1
langchain-community>=0.2.1
langchain-core>=0.2.1
sentence-transformers>=2.7.0
chromadb>=0.4.24
python-dotenv>=1.0.0
notion-client>=2.0.0
PyPDF2>=3.0.1
markdown>=3.5.2
beautifulsoup4>=4.12.3
pydantic>=2.6.1
requests>=2.32.0
werkzeug>=3.0.1
gunicorn>=21.2.0

